Rename NetworkGlobals to PeersData
