+++ packages/lodestar/src/network/reqresp/reqResp.ts
@@ -19,7 +19,7 @@ import {IPeerRpcScoreStore} from "../peers";
 import {assertSequentialBlocksInRange, formatProtocolId} from "./utils";
 import {MetadataController} from "../metadata";
 import {INetworkEventBus, NetworkEvent} from "../events";
-import {NetworkGlobals} from "../globals";
+import {PeersData} from "../peers/peersData";
 import {ReqRespHandlers} from "./handlers";
 import {IMetrics} from "../../metrics";
 import {RequestError, RequestErrorCode} from "./request";
@@ -46,7 +46,7 @@ export type IReqRespOptions = Partial<typeof timeoutOptions>;
 export class ReqResp implements IReqResp {
   private config: IBeaconConfig;
   private libp2p: LibP2p;
-  private readonly networkGlobals: NetworkGlobals;
+  private readonly peersData: PeersData;
   private logger: ILogger;
   private reqRespHandlers: ReqRespHandlers;
   private metadataController: MetadataController;
@@ -62,7 +62,7 @@ export class ReqResp implements IReqResp {
   constructor(modules: IReqRespModules, options: IReqRespOptions & RateLimiterOpts) {
     this.config = modules.config;
     this.libp2p = modules.libp2p;
-    this.networkGlobals = modules.networkGlobals;
+    this.peersData = modules.peersData;
     this.logger = modules.logger;
     this.reqRespHandlers = modules.reqRespHandlers;
     this.metadataController = modules.metadata;
@@ -153,9 +153,9 @@ export class ReqResp implements IReqResp {
     try {
       this.metrics?.reqRespOutgoingRequests.inc({method});
 
-      const encoding = this.networkGlobals.getEncodingPreference(peerId.toB58String()) ?? Encoding.SSZ_SNAPPY;
+      const encoding = this.peersData.getEncodingPreference(peerId.toB58String()) ?? Encoding.SSZ_SNAPPY;
       const result = await sendRequest<T>(
-        {forkDigestContext: this.config, logger: this.logger, libp2p: this.libp2p, networkGlobals: this.networkGlobals},
+        {forkDigestContext: this.config, logger: this.logger, libp2p: this.libp2p, peersData: this.peersData},
         peerId,
         method,
         encoding,
@@ -191,14 +191,14 @@ export class ReqResp implements IReqResp {
       // TODO: Do we really need this now that there is only one encoding?
       // Remember the prefered encoding of this peer
       if (method === Method.Status) {
-        this.networkGlobals.setEncodingPreference(peerId.toB58String(), encoding);
+        this.peersData.setEncodingPreference(peerId.toB58String(), encoding);
       }
 
       try {
         this.metrics?.reqRespIncomingRequests.inc({method});
 
         await handleRequest(
-          {config: this.config, logger: this.logger, networkGlobals: this.networkGlobals},
+          {config: this.config, logger: this.logger, peersData: this.peersData},
           this.onRequest.bind(this),
           stream,
           peerId,
