+++ packages/lodestar/test/unit/network/reqresp/response/index.test.ts
@@ -5,7 +5,7 @@ import {LodestarError} from "@chainsafe/lodestar-utils";
 import {RespStatus} from "../../../../../src/constants";
 import {Method, Encoding, Version} from "../../../../../src/network/reqresp/types";
 import {handleRequest, PerformRequestHandler} from "../../../../../src/network/reqresp/response";
-import {NetworkGlobals} from "../../../../../src/network/globals";
+import {PeersData} from "../../../../../src/network/peers/peersData";
 import {expectRejectedWithLodestarError} from "../../../../utils/errors";
 import {expectEqualByteChunks, MockLibP2pStream} from "../utils";
 import {sszSnappyPing} from "../encodingStrategies/sszSnappy/testData";
@@ -18,7 +18,7 @@ chai.use(chaiAsPromised);
 describe("network / reqresp / response / handleRequest", async () => {
   const logger = testLogger();
   const peerId = getValidPeerId();
-  const networkGlobals = new NetworkGlobals();
+  const peersData = new PeersData();
 
   let controller: AbortController;
   beforeEach(() => (controller = new AbortController()));
@@ -73,7 +73,7 @@ describe("network / reqresp / response / handleRequest", async () => {
       const stream = new MockLibP2pStream(requestChunks);
 
       const resultPromise = handleRequest(
-        {config, logger, networkGlobals},
+        {config, logger, peersData: peersData},
         performRequestHandler,
         stream,
         peerId,
